refactor: fix Games component
