+++ src/components/Games/Games.css
@@ -3,8 +3,6 @@
   justify-content: center;
   align-items: center;
   flex-wrap: wrap;
-  width: 100%;
-
 }
 
 .gamecard__block {
